Add reid_arxiv template for DV5 grating
Add a reid_arxiv template for the DV5 grating, constructed using the
`templates.build_template()` function from three different spectra spanning the
useful spectral range for this grating.

This one grating seems to confound the `holy-grail` wavelength calibration
method, so this reid_arxiv is used with `full-template` for the DV5 (500/5500)
grating for successful calibration.

	new file:   pypeit/data/arc_lines/reid_arxiv/ldt_deveny_500l_HgCdAr.fits