Fix bug in flatfield.py related to `pixelflat_min_wave`
When the `pixelflat_min_wave` and `pixelflat_max_wave` parameters were
introduced in #1225, they required the use of a `waveimg` array specifying
the wavelength at each pixel on the slit(s) in the flat frame.  The `waveimg`
was included in #1225, but placed BEFORE the slit edges are tweaked in the
`FlatField.fit()` method.

If the slit edges are tweaked OUTWARD, the `onslit_tweak` mask will then include
parts of `waveimg` outside the original slit mask (and therefore have wavelength
equal to zero).  When the `pixelflat_min_wave` parameter is invoked (even with
a value of zero), the code finds matching wavelengths in all rows of the image,
and the pixelflat_norm becomes identically one.

Because `waveimg` is only used within this method for the `pixelflat_min_wave`
and `pixelflat_max_wave` parameters, this commit moves its creation to AFTER the
slit tweaking and just before their use.

	modified:   pypeit/flatfield.py