Changes to allow for arbitrary binning
With arbitrary binning, need to actually use the TRIMSEC and BIASSEC keywords
from the FITS header.  Since these are a "backwards" from what PypeIt expects,
add new method `swap_section` to the LDTDeVenySpectrograph class to parse and
return the keywords in the correct form.

It is possible that `swap_section` is applicable more generally and should move
to the pypeit.core.parse module, but I will leave that decision to the PypeIt
development team.

Clean up the `binning` compound_meta case to use pypeit.core.parse functions
exclusively.

Finally, if arbitrary binning is allowed, it must be added as one of the
configuration keys for defining a unique instrument setup.

	modified:   pypeit/spectrographs/ldt_deveny.py