Add `get_rawimage()` method to LDTDeVenySpectrograph
Add a specialized method to override the base Spectrograph class method for
reading in and parsing the raw images for LDT/DeVeny.

The need for a DeVeny-specific `get_rawimage()` method was not apparent until
PypeIt was tested with binned data (2x1, spatial x spectral).  The base class
method assumes the DATASEC and OSCANSEC regions supplied are for the unbinned
data, and it adjusts the trim accordingly.  The LOIS control system for
LDT/DeVeny automatically adjusts the FITS header values (which are fed into
PypeIt) based on the binning, so this caused excessive trim of the binned
images.

The present commit is a stripped-down version of the base class method, being
careful to NOT send the binning to the parser.

	modified:   pypeit/spectrographs/ldt_deveny.py