Reverting datamodel version
The inclusion of `minimum_version` in `pypeit.flatfield.FlatImages()` seemed to
indicate that multiple versions of a data model would be valid, so long as they
lie between `minimim_version` and `version`, inclusive.  Unfortunately, it
appears that this functionality has not yet been implemented.

The version check happens in `pypeit.datamodel.DataContainer._parse()`, and only
includes equality statements.  Additionally, the class attribute
`minimum_version` is not referenced in the `_parse()` method and only appears
in the active code base in the data container classes of `pypeit/alignframe.py`,
`pypeit/flatfield.py`, and `pypeit/slittrace.py`.

This leads to a quandary of how to proceed.  This commit retains the original
data model version `1.1.0`, despite the fact that the model has changed
slightly.  We could update the version number (and update the Cooked data), but
this may cause unnecessary disruption for users upon upgrade.

The final option would be to modify `DataContainer._parse()` to check the data
container version against both the current and minumum versions.  This would
seem like the most correct path forward, but would require extensive testing
and may be outside the scope of this particular PR.

	modified:   pypeit/flatfield.py