Refactor the `tellgridfile` parameter
Previously, the Telluric Atmospheric Grids, as specified in the parameter
par['sensfunc']['IR']['telgridfile'], were specified by full path.  This is
different from, say, the `reid_arxiv` files, which are specified by filename
only with the loader supplying the path within the PypeIt directory structure.

This commit shifts the `telgridfile` parameter to match the usage of other data
files within PypeIt.  The Telluric Grid path is now solely defined in
`pypeit/par/pypeitpar.py`, and every place that calls the Grids goes back to
this base parameter definition.  In particular, the main loader for the
Telluric Grids (`pypeit.core.telluric.read_telluric_grid()`) joins the path
with the specific filename desired.

The goal for this change is to make it easier to specify a Telluric Grid in
a Parameter Block of a `.sens` file, or for users of a spectrograph that doesn't
have a default Telluric Grid to easily try out a grid with their data.

Documentation changes will be needed to ensure both everyday users and power
users know how to interact with this change.

	modified:   pypeit/core/telluric.py
	modified:   pypeit/par/pypeitpar.py
	modified:   pypeit/scripts/install_telluric.py
	modified:   pypeit/scripts/sensfunc.py
	modified:   pypeit/scripts/tellfit.py
	modified:   pypeit/spectrographs/gemini_flamingos.py
	modified:   pypeit/spectrographs/gemini_gmos.py
	modified:   pypeit/spectrographs/gemini_gnirs.py
	modified:   pypeit/spectrographs/keck_deimos.py
	modified:   pypeit/spectrographs/keck_mosfire.py
	modified:   pypeit/spectrographs/keck_nires.py
	modified:   pypeit/spectrographs/keck_nirspec.py
	modified:   pypeit/spectrographs/magellan_fire.py
	modified:   pypeit/spectrographs/mmt_binospec.py
	modified:   pypeit/spectrographs/mmt_mmirs.py
	modified:   pypeit/spectrographs/p200_dbsp.py
	modified:   pypeit/spectrographs/p200_tspec.py
	modified:   pypeit/spectrographs/shane_kast.py
	modified:   pypeit/spectrographs/soar_goodman.py
	modified:   pypeit/spectrographs/vlt_sinfoni.py
	modified:   pypeit/spectrographs/vlt_xshooter.py
	modified:   pypeit/tests/tstutils.py