Silence VerifyWarning for "Keyword name ... is greater than 8 characters"
This is a particularly annoying AstroPy Warning, and one that crops up
frequently in a PypeIt reduction because much of the added metadata generated
during the reduction is added to the header with a keyword > 8 characters in
length.

Because the PypeIt team chose to use long keywords for its metadata, it does not
seem helpful for end users to see a flood of such messages during their
reductions.  In addition, these warnings break up the flow of messages needed by
developers for troubleshooting.

NOTE: It could be possible to pass down an option to not ignore the Warning.
The places that call `pypeit.io.write_to_hdu` are `pypeit.io.write_to_fits` and
`pypeit.datamodel.DataContainer.to_hdu`.  This would be for developers for whom
the HIERARCH keyword is problematic.

	modified:   pypeit/io.py