First steps toward keeping package data remotely
To reduce the package size on PyPI, keep package data files (such as
`reid_arxiv`) remotely.  This branch will introduce code to check on the local
machine for a requested file, and download it from the remote server if not
present.

As most users only need a small number of `reid_arxiv` files, this keeps things
lean.

This commit simply moves the construction of the full path for the `reid_arxiv`
to `pypeit.data.utils`.  More to come...

	modified:   pypeit/core/wavecal/waveio.py
	modified:   pypeit/data/utils.py
	modified:   pypeit/scripts/show_arxiv.py