Download/Cache for reid_arxiv, sensfunc, and skisim files
Create path-returning functions for reid_arxiv, sensfunc, and skisim files
from the `pypeit/data` directory.

If the file is in the package directory, functions return that path.  This
should only be the case for from-source installations (i.e., developers).
For all end-user installs (e.g., from `pip` or `conda`), these files will not
be included in the distribution for compactness of the package.  For these
users, use AstroPy's `download_file` functionality to download and cache these
files (currently from the GitHub repo) for use.  By default, the cache is in
the user's home directory -- but can be configured through AstroPy's interface.

Next steps will include documentation changes to describe how all this works.

	modified:   MANIFEST.in
	modified:   pypeit/core/coadd.py
	modified:   pypeit/core/flux_calib.py
	modified:   pypeit/core/wavecal/waveio.py
	modified:   pypeit/data/utils.py
	modified:   pypeit/sensfilearchive.py
	modified:   pypeit/wavemodel.py