Overhaul of `LDT/DeVeny` wavelength templates
Based on recent data, this commit adds a new set of wavelength calibration
templates for the 150g/mm, 300g/mm, 600g/mm, and 1200g/mm gratings.  In the
process, several new lines have been added to the HgI and CdI linelists.  New
arc lines in the lists from DeVeny are now listed with the template in which
it was identified.

The template-creation routines are being added here, and the input WaveCalib
files will be submitted to the `dev-suite` in a PR there.

Finally, adjustments to the configuration-specific parameters were made to
conform to the new templates.

	new file:   pypeit/core/wavecal/spectrographs/templ_ldt_deveny.py
	modified:   pypeit/data/arc_lines/lists/CdI_lines.dat
	modified:   pypeit/data/arc_lines/lists/HgI_lines.dat
	new file:   pypeit/data/arc_lines/reid_arxiv/ldt_deveny_1200_HgCdAr.fits
	new file:   pypeit/data/arc_lines/reid_arxiv/ldt_deveny_150_HgCdAr.fits
	deleted:    pypeit/data/arc_lines/reid_arxiv/ldt_deveny_150l_HgCdNeAr.fits
	new file:   pypeit/data/arc_lines/reid_arxiv/ldt_deveny_300_HgCdAr.fits
	deleted:    pypeit/data/arc_lines/reid_arxiv/ldt_deveny_300l_HgCdAr.fits
	renamed:    pypeit/data/arc_lines/reid_arxiv/ldt_deveny_500l_HgCdAr.fits -> pypeit/data/arc_lines/reid_arxiv/ldt_deveny_500_HgCdAr.fits
	new file:   pypeit/data/arc_lines/reid_arxiv/ldt_deveny_600_HgCdAr.fits
	deleted:    pypeit/data/arc_lines/reid_arxiv/ldt_deveny_600l_HgCdNeAr.fits
	modified:   pypeit/spectrographs/ldt_deveny.py