Add ability to "pre-cache" needed GitHub-hosted package data
At the request of @profxj, this commit adds the ability to pre-cache GitHub-
hosted package data in the event that a user wants to use PypeIt while not
connected to the internet.

Adds a new console script `pypeit_cache_github_data` that takes the name of
a spectrograph and caches the necessary package data from GitHub.  This may be
done rather than downloading/caching on the fly during a PypeIt run.

In order to quickly and easily find the necessary files on GitHub, this commit
adds a new dependency for PypeIt: `pygithub`

	modified:   environment.yml
	modified:   pypeit/data/utils.py
	new file:   pypeit/scripts/cache_github_data.py
	modified:   setup.cfg