Add ability to search the PypeIt cache for arc line lists
Infrastructure side of things to allow for users to call their own line lists
for wavelength calibration.  This commit adds a new `data.get_linelist_filepath()`
function, similar to other `get_*_filepath()` functions.  This case is slightly
different because the base linelists are still distributed with the ``pip``
package, but the functionality pulls user files from the cache, if they exist.

Next steps include building the install script for user line lists, and updating
the documentation to describe how to access this functionality.

	modified:   pypeit/core/wavecal/kdtree_generator.py
	modified:   pypeit/core/wavecal/templates.py
	modified:   pypeit/core/wavecal/waveio.py
	modified:   pypeit/data/arc_lines/convert_NIST_to_lists.py
	modified:   pypeit/data/utils.py
	new file:   pypeit/deprecated/waveio_old.py