(HOTFIX) Fix typo related to GitHub download
Fixes a typo in the constructed URL of GitHub-hosted content, where a second
(spurious) "/" is included in the URL.  When AstroPy goes to download the file,
the extra slash is removed and the correct file is downloaded.  However, when
the code attempts to locate the file, it cannot find it in the cache because of
the slash mismatch.  For most users, the file is re-downloaded.  For offline
uses (for instance, compute clusters), this causes a fatal error.

	modified:   pypeit/data/utils.py