Update internal data paths to use `pathlib`
Shift from `os.path.join()` usage to `pathlib` for all internal data file
locations.  This allows for all of the functionality of `pathlib` objects and
allows the discontinuance of `import os` in most cases.

Enforce `python>=3.9` in `setup.cfg` and `environment.yml`, which allows the
simplification of the `data.Paths` class to use the dual decorators of
`@classmethod` and `@property` without having to resort to a metaclass
structure.

Various modernization of code happened upon in the process of making the above
changes, along with cleaning up various `pytest` warnings.

	modified:   environment.yml
	modified:   pypeit/core/convert_DEIMOSsavfiles.py
	modified:   pypeit/core/extract.py
	modified:   pypeit/core/findobj_skymask.py
	modified:   pypeit/core/flat.py
	modified:   pypeit/core/flexure.py
	modified:   pypeit/core/flux_calib.py
	modified:   pypeit/core/procimg.py
	modified:   pypeit/core/skysub.py
	modified:   pypeit/core/telluric.py
	modified:   pypeit/core/wavecal/autoid.py
	modified:   pypeit/core/wavecal/patterns.py
	modified:   pypeit/core/wavecal/templates.py
	modified:   pypeit/core/wavecal/waveio.py
	modified:   pypeit/data/arc_lines/convert_NIST_to_lists.py
	modified:   pypeit/data/utils.py
	modified:   pypeit/scripts/ql_multislit.py
	modified:   pypeit/spectrographs/keck_deimos.py
	modified:   pypeit/spectrographs/keck_hires.py
	modified:   pypeit/spectrographs/keck_lris.py
	modified:   pypeit/spectrographs/vlt_xshooter.py
	modified:   pypeit/tests/test_flexure.py
	modified:   pypeit/tests/test_pkgdata.py
	modified:   pypeit/tests/test_runpypeit.py
	modified:   pypeit/tests/tstutils.py
	modified:   pypeit/utils.py
	modified:   setup.cfg