Allow for configuration-specific parameters in `pypeit_sensfunc`
Add the means for allowing the use of configuration-specific parameters from the
Spectrograph subclasses in the sensitivity function calculation.

Switch from using the static `spectrograph.default_pypeit_par()` method to the
dynamic `spectrograph.config_specific_par()`.  The issue is that this latter
method uses the FITS header of the raw file to compute compound metadata keys,
keys that are sometimes absent in the `spec1d` files.  This commit allows for
the specification of FITS cards from the raw data file to be included in
downstream output files within PypeIt, thus allowing for configuration-specific
parameters in the "afterburner" scripts.

	modified:   pypeit/core/flux_calib.py
	modified:   pypeit/scripts/sensfunc.py
	modified:   pypeit/sensfunc.py
	modified:   pypeit/spectrographs/spectrograph.py