Propagate the `raw_header_cards()` method
Instead of just being handy, the `raw_header_cards()` method to the
Spectrograph class is actually necessary for all spectrographs that have a
`config_specific_par()` method that uses PypeIt metadata keys.

This commit adds that necessary method to all classes, and does some slight
reordering of the methods in some classes to keep these 4 methods in the same
order:  `init_meta()`, `compund_meta()`, `configuration_keys()`, and
`raw_header_cards()`.  The reason for this is to easily find the necessary
raw FITS keywords that correspond to the needed PypeIt keys.