Collection of sundry changes:
* Allow `pypeit_chk_wavecalib` to accept lists of input files

If more than one WaveCalib is to be checked at once, this commit allows for
the script to accept a list of input files, and simply loops the main execution
over the list of input files.

* Allow restricted wavelength range for flexure cross-correlation

Add FlexurePar parameters ``minwave`` and ``maxwave`` that may be used to
restrict the wavelength range used for the flexure cross-correlation.  This is
useful in the cases where the wavelength calibration _at the edges_ of the
spectral range is poorly constrained, yet there is a night-sky line used in
the correlation there.

Other uses may be possible.

The new parameters are passed to the wavelength restrictor in
`pypeit.core.flexure.spec_flux_shift` as 0 and numpy.inf as the effective min
and max specified values, respectively, (while the actual parameters are
``None`` if not specified).

* Tweak Hg wavelength

* Adjust wavecal fit parameters for LDT/DeVeny

* Update verbosity assignment in show_2dspec

	modified:   pypeit/core/flexure.py
	modified:   pypeit/data/arc_lines/lists/Hg_DeVeny1200_lines.dat
	modified:   pypeit/extraction.py
	modified:   pypeit/find_objects.py
	modified:   pypeit/par/pypeitpar.py
	modified:   pypeit/scripts/chk_wavecalib.py
	modified:   pypeit/scripts/show_2dspec.py
	modified:   pypeit/spectrographs/ldt_deveny.py