Add optional metadata table validation method
To deal with instrument-specific software issues at the telescope,
LDT/NIHTS needs a method for changing metadata table values to match
temporally adjacent frames.

This commit adds a new Spectrograph method called "validate_fitstbl()"
that can be used for this purpose.  The base class implementation of
this method simply returns the metadata table unchanged.  As such, all
spectrographs that do not implement this new method will not be affected
by the addition.  The new method is called from within `pypeitsetup.py`
as part of building the metadata table in the first place.

	modified:   pypeit/pypeitsetup.py
	modified:   pypeit/spectrographs/spectrograph.py