Fix metadata setting of frame types when input array is empty
If the input array of ``type_bits`` to
``pypeit.metadata.PypeItMetaData.frame_paths()`` is empty, the
resulting Column data type is returned as a float64.  Since the
self-proclaimed kludge checking of dtype is looking for unicode
strings, this causes an unusual error.  By forcing the Column
to have a string data type, the kludge checking works fine and
life goes on.

	modified:   pypeit/metadata.py