``numpydev`` tests not testing bleeding-edge NumPy
The ``numpydev`` tests currently test against the latest released
version of NumPy (on 20-Oct-2025, that is ``numpy==2.3.4``), but with
this change, the ``numpydev`` tests will actually download the latest
nightly wheel (on 20-Oct-2025, that is ``numpy==2.4.0.dev0``).

We can discuss whether this is what we actually want to be doing, but
this commit aligns the name of the test with actual functionality.

	modified:   tox.ini